improve

+++ tools/nutool/nutool.Test/CommonTestData.cs
@@ -6,31 +6,45 @@ public static class CommonTestData
     internal const string _PathTestAssets = "TestAssets";
 
     internal static string _TempPath = Path.GetTempPath();
-    internal static string _TempDir = Path.Combine(_TempPath, "nutool.Test");
+    internal static string _TempDir = JoinPath(_TempPath, "nutool.Test");
 
-    private static string GetPath(string path) => Path.Combine(_TempDir, path);
+    private static string GetPath(string path) => JoinPath(_TempDir, path);
 
     internal static string PathTestOutput => GetPath(_PathTestOutput);
     internal static string PathTestAssets => GetPath(_PathTestAssets);
 
     internal static string GetScopedOutputDirectoryPath(string name)
     {
-        var path = Path.Combine(PathTestOutput, name);
+        var path = JoinPath(PathTestOutput, name);
         Directory.CreateDirectory(path);
         return path;
     }
 
+    internal static string JoinPath(params string[] paths) => Path.GetFullPath(Path.Combine(paths));
+
+    private static readonly Lock _LockObject = new();
+    private static bool _Initialized;
+
     internal static void Init()
     {
-        if (Directory.Exists(_TempDir))
+        lock (_LockObject)
         {
-            Directory.Delete(_TempDir, true);
-        }
-        Directory.CreateDirectory(_TempDir);
-        CopyFilesRecursively("TestAssets", PathTestAssets);
+            if (_Initialized)
+            {
+                return;
+            }
+            _Initialized = true;
 
-        Ui.Out = new StringWriter();
-        Ui.Err = new StringWriter();
+            if (Directory.Exists(_TempDir))
+            {
+                Directory.Delete(_TempDir, true);
+            }
+            Directory.CreateDirectory(_TempDir);
+            CopyFilesRecursively("TestAssets", PathTestAssets);
+
+            Ui.Out = new StringWriter();
+            Ui.Err = new StringWriter();
+        }
     }
 
     internal static void CopyFilesRecursively(string sourcePath, string targetPath)
